vllm-delivery-template/pyproject.toml
[project]
name = "vllm-delivery-template"
version = "0.1.0"
description = "Production-ready vLLM delivery template with OpenAI-compatible API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "vllm>=0.6.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["serving", "client", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
